Rename src/model/clique_diff.py to glom_diff.py
Two unrelated modules were named clique_diff: this one, a library that
diffs glom-state clique sets for source-impact-report, and
tools/clique_diff/diff.py, the build-vs-build CLI. Name each for what it
diffs. glom_diff takes the union-find output of glom(); the next commit
introduces compendium_diff, which takes finished JSONL compendia.

Pure rename plus import and docstring updates.
